solaris-browse/pyproject.toml
[tool.poetry]
name = "quickstart-selenium-python"
version = "0.1.0"
description = ""
authors = ["Anirudh Kamath <anirudh@browserbase.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
selenium = ">=4.26.1"
browserbase = "1"
python-dotenv = "1.0.0"
anthropic = "^0.39.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
